login-ms/requirements.txt
fastapi
uvicorn[standard]
pydantic
passlib
sqlalchemy
pyjwt